2010-04-12  Basile Starynkevitch  <basile@starynkevitch.net> 	{{bug reported by Sylvestre Ledru}} 	* melt-runtime.c (melt_debug_out): Added missing cast from 	unsigned short to int.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/melt-branch@158244 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/melt-runtime.c
@@ -9112,7 +9112,7 @@ melt_debug_out (struct debugprint_melt_st *dp,
 	  {
 	    fputs ("[", dp->dfil);
 	    if (p->obj_vartab)
-	      for (ix = 0; ix < p->obj_len; ix++)
+	      for (ix = 0; ix < (int) p->obj_len; ix++)
 		{
 		  if (ix > 0)
 		    skip_debug_out (dp, depth);
